nexus-adk-agent/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
google-adk>=0.0.1
google-genai>=0.1.0
